% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_file.R
\name{clear_all}
\alias{clear_all}
\title{clear all files in a directory}
\usage{
clear_all(clear_dir, file_ext = NULL)
}
\arguments{
\item{clear_dir}{string indicating the directory of files to clear}

\item{file_ext}{character vector of file extensions of files to remove}
}
\description{
clear all files in a directory
}
