% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_file.R
\name{clear_files}
\alias{clear_files}
\title{clear files}
\usage{
clear_files(clear_dir, file_name)
}
\arguments{
\item{clear_dir}{string indicating the directory of files to clear}

\item{file_name}{character vector of file names to remove}
}
\description{
clear files
}
