% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_file.R
\name{copy_file}
\alias{copy_file}
\title{copy a file}
\usage{
copy_file(source_file, dest_dir, overwrite, timestamp)
}
\arguments{
\item{source_file}{string with file path  of the object to copy}

\item{dest_dir}{string indicating the directory to paste copied file to}

\item{overwrite}{logical indicating whether existing files in destination directory should be overwritten}

\item{timestamp}{logical indicating whether you'd like to timestamp the copied file}
}
\description{
copy a file
}
