% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_file.R
\name{copy_files}
\alias{copy_files}
\title{copy multiple files from a windows directory}
\usage{
copy_files(source_files, dest_dir, overwrite = FALSE, timestamp = FALSE)
}
\arguments{
\item{source_files}{\code{data.frame} with file_path and file_name of the object to copy}

\item{dest_dir}{string indicating the directory to paste copied file to}

\item{overwrite}{logical indicating whether existing files in destination directory should be overwritten - default FALSE results in no copy}

\item{timestamp}{logical indicating whether you'd like to timestamp the copied file}
}
\description{
copy multiple files from a windows directory
}
