% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_results_folder.R
\name{create_results_folder}
\alias{create_results_folder}
\title{Creates a folder in the 'results' directory.}
\usage{
create_results_folder(folder_name)
}
\arguments{
\item{folder_name}{Character input. Sub folder name.}
}
\description{
Checks for the existence of specified sub folder in 'results' directory and
creates it if it does not exist. Requires 'data' and 'results' directory to be
set up.
}
