% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.Files.R
\name{extract.Files}
\alias{extract.Files}
\title{Extract files based on user input to select the input and output directory, and to select the file type}
\usage{
extract.Files(path = getwd(),
                      exdir = getwd(),
                      file.type = c('.zip','.gz'),
                      overwrite.files = F)
}
\arguments{
\item{path}{This is the path for the input directory. Leaving this blank will set the current working directory as the
input directory. Typing in "select" as the path will allow you to select the directory based on user input.}

\item{exdir}{This is the directory to be used as the output for extracting the files. Leaving this blank will set the
current working directory as the output directory. Typing in "select" as the path will allow you to select the directory
based on user input.}

\item{overwrite.files}{This determines whether or not files with the same name will be overwritten in the output directory.}
}
\description{
\code{extract.Files} uses R.utils and utils to extract files from .zip and .gz files. It takes user input in order to
select the input and output directories. Any other file types will be ignored.
}
