% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_names}
\alias{file_names}
\title{File Names}
\usage{
file_names(file = "file", n = 1L, ext = "rds")
}
\arguments{
\item{file}{A string of the file name.}

\item{n}{A count of the number of files.}

\item{ext}{A string of the extension.}
}
\value{
A character vector of the file names.
}
\description{
File Names
}
\examples{
file_names(n = 10)
}
