% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileR-package.R
\docType{package}
\name{fileR-package}
\alias{fileR}
\alias{fileR-package}
\title{fileR: A pipeable interface to directory and folder creation and use}
\description{
Pipeable create and save operations to directories or 'folders'. This package hopes increase organization of files within R projects by providing a helper functions to help organize inputs and outputs (e.g., save result as csv to folder).
}
\keyword{internal}
