% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filer.R
\name{filer}
\alias{filer}
\title{Get file path transparently for a possibly-gzip-compressed file, with lots of useful checks}
\usage{
filer(name, ext = "txt", gzip = FALSE, exists = FALSE,
  logical = FALSE)
}
\arguments{
\item{name}{Base name of file (mandatory).}

\item{ext}{Extension (default txt).  Period between \code{name} and \code{ext} gets added automatically unless \code{ext==''}.}

\item{gzip}{If true, adds .gz extension after usual extension.}

\item{exists}{If true, checks that file exists; if file did not exist, checks if compressed version exists (returns that path in that case).  If \code{gzip=TRUE}, then only the compressed version is checked for existence.}

\item{logical}{If true, function will return a boolean for whether file exists or not (regular or compressed).}
}
\value{
If \code{exists=FALSE} and \code{logical=FALSE}, returns the desired relative file path, with extensions added as requested (no existence tests).
If \code{exists=TRUE}, returns the path of the file that exists (adds .gz extension if only that file exists, even if \code{gzip=FALSE}), stopping (fatally) if the file is missing.
If \code{logical=TRUE}, returns \code{TRUE} if file (regular or compressed) exists, \code{FALSE} otherwise.
}
\description{
Returns a (relative but otherwise complete) file path from its base name and optional extensions (default .txt).
The main goal is to handle compressed files transparently, without having to know ahead of time if the compressed or uncompressed version is the one that is present.
If \code{gzip=TRUE}, .gz extension gets added.
If \code{exists=TRUE}, existing uncompressed or compressed file path is returned, or stops if file is missing.
If \code{logical=TRUE}, returns a boolean that indicates if file exists or not, testing both uncompressed and compressed versions.
}
\details{
Behavior when multiple modifiers are \code{TRUE}:
\code{logical=TRUE} overrides \code{exists=TRUE} (returns boolean, non-fatal).
\code{gzip=TRUE} restricts file existence test to compressed file when either \code{logical=TRUE} or \code{exists=TRUE}.
}
\examples{
# examples that simply construct file path (without existence tests)
filer('myfile') # returns "myfile.txt"
filer('myfile', 'pdf') # returns "myfile.pdf"
filer('myfile', '') # returns "myfile" (no extension)
filer('myfile', gzip=TRUE) # returns "myfile.txt.gz"
filer('myfile', 'yml', gzip=TRUE) # returns "myfile.yml.gz"
filer('myfile', '', gzip=TRUE) # returns "myfile.gz"
# no special treatment for periods in name (not assumed to be existing extensions)
filer('myfile.blah') # returns "myfile.blah.txt"
filer('myfile.blah', '') # returns "myfile.blah" (no added extension)

# test that file exists
# returns TRUE if myfile.txt or myfile.txt.gz exist, FALSE otherwise
filer('myfile', logical=TRUE)
# returns TRUE if myfile.txt.gz exist, FALSE otherwise (even if myfile.txt exists)
filer('myfile', gzip=TRUE, logical=TRUE)

\dontrun{
### (these examples are fatal when the file in question doesn't exist)

# returns myfile.txt if it exists,
# or myfile.txt.gz if that exist,
# or stops with an informative message if file does not exist
filer('myfile', exists=TRUE)

# returns myfile.txt.gz if it exist,
# or stops with an informative message if file does not exist
# (even if myfile.txt exists)
filer('myfile', gzip=TRUE, exists=TRUE)
}

}
