% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_files.R
\name{list_data_files}
\alias{list_data_files}
\title{List files in data directory}
\usage{
list_data_files(folder_name = NULL)
}
\arguments{
\item{folder_name}{Character input. 'data' sub folder name.}
}
\value{
List of file names.
}
\description{
Searches data directory for optional sub folder denoted by folder_name
argument. If folder does not exist, function will throw an error.
}
