% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_all_files_to_results_folder.R
\name{move_all_files_to_results_folder}
\alias{move_all_files_to_results_folder}
\title{Recursively moves all files in the results directory except those in the target
folder to the target folder.}
\usage{
move_all_files_to_results_folder(to)
}
\arguments{
\item{to}{Character input. Specify sub folder in 'results' directory to read
from.}
}
\description{
Copies and pastes all files in the results 'folder' to a target folder while
ignoring the target folder. Original files are then deleted.
}
