% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_files_between_results_folders.R
\name{move_files_between_results_folders}
\alias{move_files_between_results_folders}
\title{Moves files from one 'folder' in the results directory to another 'folder'.}
\usage{
move_files_between_results_folders(from, to)
}
\arguments{
\item{from}{Character input. Specify 'folder' in results directory to move files
from.}

\item{to}{Character input. Specify 'folder' in results directory to move files
to.}
}
\description{
Copies and pastes files into target results 'folder.' Original files are then
deleted.
}
