% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_csv_from_data.R
\name{read_all_csv_from_data}
\alias{read_all_csv_from_data}
\title{Retrieves all files from the 'data' directory as a list.}
\usage{
read_all_csv_from_data(folder_name = NULL, ...)
}
\arguments{
\item{folder_name}{Optional. Character input. Specify sub folder in 'data'
directory to read from.}

\item{...}{Passes arguments to `read.csv()`.}
}
\description{
Reads all csv files in the 'data' directory or in a sub folder of the 'data'
directory and saves them all as a list. File names are used as names for
list items to allow easy identification of files.
}
