% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_from_data.R
\name{read_csv_from_data}
\alias{read_csv_from_data}
\title{Retrieve file from data directory}
\usage{
read_csv_from_data(file_name, folder_name = NULL, ...)
}
\arguments{
\item{file_name}{Character input. Name of the file to be retrieved from
'data' directory without file extension.}

\item{folder_name}{Optional. Character input. Specify sub folder in 'data'
directory to read from.}

\item{...}{Passes arguments to `read.csv()`.}
}
\value{
File contents as an object.
}
\description{
A wrapper around `read.csv()` but with a pre-determined path to the data
directory. Sub folder is optional in data directory. If a folder is
specified but does not currently exist, the function will create that
folder.
}
