% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_from_results.R
\name{read_csv_from_results}
\alias{read_csv_from_results}
\title{Retrieve file from results directory}
\usage{
read_csv_from_results(file_name, folder_name = NULL, ...)
}
\arguments{
\item{file_name}{Character input. Name of the file to be retrieved from
'results' directory without file extension.}

\item{folder_name}{Optional. Character input. Specify sub folder in 'results'
directory to read from.}

\item{...}{Passes arguments to `read.csv()`.}
}
\value{
File contents as an object.
}
\description{
A wrapper around `read.csv()` but with a pre-determined path to the results
directory. Sub folder is optional in results directory.
}
