% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_csv_to_results_folder.R
\name{save_csv_to_results_folder}
\alias{save_csv_to_results_folder}
\title{Pipeable function that saves the object passed in as a csv in the 'results'
directory.}
\usage{
save_csv_to_results_folder(x, folder_name, file_name, ...)
}
\arguments{
\item{x}{Object to be saved and invisibly returned.}

\item{folder_name}{Character input. 'results' sub folder name.}

\item{file_name}{Character input. Name of the csv file without the file
extension.}

\item{...}{Passes arguments to `write.csv()`}
}
\value{
The object 'x' that was passed in invisibly.
}
\description{
A wrapper around `write.csv()` but with a pre-determined path to
a folder in the 'results' directory. Will force sub folder path/creation in
'results' directory.
}
