% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.WorkingDirectory.R
\name{select.WorkingDirectory}
\alias{select.WorkingDirectory}
\title{Set working directory based on user input.}
\usage{
select.WorkingDirectory()
}
\description{
\code{select.WorkingDirectory} allows the user to set the working directory based on their input. The are no arguments
to this function. Once you run it, you select the working directory with input placed in the console.
}
