% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_files}
\alias{write_files}
\title{Write Files}
\usage{
write_files(.objects, .dir = ".", .fun = saveRDS,
  .ask = getOption("filer.ask", TRUE), .silent = FALSE, ...)
}
\arguments{
\item{.objects}{A uniquely named list of objects to save as files.}

\item{.dir}{A string of the path to the directory.}

\item{.fun}{The function to use when saving each object.}

\item{.ask}{A flag indicating whether to ask before}

\item{.silent}{A flag indicating whether error messages should be 
suppressed when trying to write a file.}

\item{...}{Additional arguments passed to .fun}
}
\value{
An invisible character vector of the names of the successfully saved files.
}
\description{
Write Files
}
\examples{
objects <- list(mtcars = datasets::mtcars, chickwts = datasets::chickwts)
write_files(objects, tempdir())
}
\seealso{
\code{\link{read_files}} and \code{\link{file_names}}
}
