% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filer-package.R
\docType{package}
\name{filer-package}
\alias{filer}
\alias{filer-package}
\title{filer: Handle Gzip'd Files Transparently}
\description{
Provides functions for easily constructing paths and querying for files that may or may not be compressed (with .gz extension).
}
\author{
\strong{Maintainer}: Alejandro Ochoa \email{alejandro.ochoa@duke.edu}

}
\keyword{internal}
